refactor(agent): extract low-hanging UI helpers
