+++ package.json
@@ -14,6 +14,7 @@
     "lint:invariants": "pnpm --dir packages/agent run lint:invariants && pnpm lint:workspace-plugin-invariants",
     "lint:workspace-plugin-invariants": "pnpm --filter @hachej/boring-workspace run lint:plugin-invariants",
     "audit:imports": "pnpm tsx scripts/audit-imports.ts",
+    "check:generated-artifacts": "pnpm tsx scripts/check-generated-artifacts.ts",
     "check:agent-isolation": "pnpm --filter @hachej/boring-agent run check:isolation",
     "check:bundle-size": "pnpm --filter @hachej/boring-workspace run check:bundle-size",
     "check:core-bundle-size": "pnpm --filter @hachej/boring-core run check:bundle-size",
@@ -21,7 +22,7 @@
     "ci:core-bundle-size": "pnpm --filter @hachej/boring-core run build && pnpm check:core-bundle-size",
     "test": "pnpm -r run test",
     "e2e": "pnpm --filter @hachej/boring-agent run test:e2e",
-    "ci": "pnpm lint && pnpm typecheck && pnpm test && pnpm lint:invariants && pnpm e2e",
+    "ci": "pnpm lint && pnpm check:generated-artifacts && pnpm typecheck && pnpm test && pnpm lint:invariants && pnpm e2e",
     "clean": "pnpm -r run clean",
     "version:patch": "node scripts/version.mjs patch",
     "version:minor": "node scripts/version.mjs minor",
